Fix workflow observability CI gates
Group workflow failure details to keep the database update API below
Clippy's argument limit, and move the relay GET helper into its own module
to preserve the Desktop source-size ratchet.
