+++ crates/buzz-workflow/src/lib.rs
@@ -242,8 +242,10 @@ impl WorkflowEngine {
                             RunStatus::Failed,
                             step_count,
                             &trace_json,
-                            Some("approval_not_supported"),
-                            Some("approval gates not yet implemented — see WF-08"),
+                            Some(buzz_db::workflow::WorkflowRunFailure {
+                                code: "approval_not_supported",
+                                message: "approval gates not yet implemented — see WF-08",
+                            }),
                         )
                         .await
                     {
@@ -262,7 +264,6 @@ impl WorkflowEngine {
                             RunStatus::Completed,
                             step_count,
                             &trace_json,
-                            None,
                             None,
                         )
                         .await
@@ -287,8 +288,10 @@ impl WorkflowEngine {
                         RunStatus::Failed,
                         progress.step_index as i32,
                         &trace_json,
-                        Some(e.code()),
-                        Some(&e.to_string()),
+                        Some(buzz_db::workflow::WorkflowRunFailure {
+                            code: e.code(),
+                            message: &e.to_string(),
+                        }),
                     )
                     .await
                 {
